arm64: Cleanup SCTLR flags
We currently have macros defining flags for the arm64 sctlr registers in both
kvm_arm.h and sysreg.h.  To clean things up and simplify move the definitions
of the SCTLR_EL2 flags from kvm_arm.h to sysreg.h, rename any SCTLR_EL1 or
SCTLR_EL2 flags that are common to both registers to be SCTLR_ELx, with 'x'
indicating a common flag, and fixup all files to include the proper header or
to use the new macro names.

Signed-off-by: Geoff Levand <geoff at infradead.org>